[PyTorch] Replace cached thread_locals with stack allocation in AOTI
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112116

This changes cached thread_local tensors to stack-allocated buffers. Since we were incidentally caching output in a thread_local, I had to add manual thread_local caching of outputs, which I implemented by caching a buffer and a Tensor whose storage is that buffer and then just memcpying the result into the cached buffer every time. Ideally, memory planning would be able to identify allocations that are the backing storage for outputs, but this should be good enough in the absence of planning.
ghstack-source-id: 206790215
@exported-using-ghexport

Differential Revision: [D50416438](https://our.internmc.facebook.com/intern/diff/D50416438/)